backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
pillow==10.0.1
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3
sqlalchemy==2.0.23
sqlitebiter==0.14.5
python-dotenv==1.0.0
cors==1.0.1
corsheaders==4.3.0
